+++ src/parser/enrichers/spell/_module.mjs
@@ -111,6 +111,7 @@ export { default as MassSuggestion } from "./MassSuggestion.mjs";
 export { default as MindBlank } from "./MindBlank.mjs";
 export { default as MinorIllusion } from "./MinorIllusion.mjs";
 export { default as MirrorImage } from "./MirrorImage.mjs";
+export { default as Mislead } from "./Mislead.mjs";
 export { default as MistyStep } from "./MistyStep.mjs";
 export { default as PassWithoutTrace } from "./PassWithoutTrace.mjs";
 export { default as Polymorph } from "./Polymorph.mjs";